+++ beacon-chain/rpc/eth/beacon/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "server.go",
         "state.go",
         "structs.go",
-        "structs_block_responses.go",
         "sync_committee.go",
         "validator.go",
     ],
@@ -119,6 +118,7 @@ go_test(
         "//beacon-chain/rpc/apimiddleware:go_default_library",
         "//beacon-chain/rpc/eth/helpers:go_default_library",
         "//beacon-chain/rpc/eth/shared:go_default_library",
+        "//beacon-chain/rpc/eth/shared/testing:go_default_library",
         "//beacon-chain/rpc/lookup:go_default_library",
         "//beacon-chain/rpc/prysm/v1alpha1/validator:go_default_library",
         "//beacon-chain/rpc/testutil:go_default_library",
