+++ beacon-chain/rpc/eth/shared/testing/BUILD.bazel
@@ -1,0 +1,10 @@
+load("@prysm//tools/go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    testonly = True,
+    srcs = ["json.go"],
+    importpath = "github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/shared/testing",
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_ethereum_go_ethereum//common/hexutil:go_default_library"],
+)
