+++ beacon-chain/rpc/eth/validator/structs.go
@@ -1,6 +1,10 @@
 package validator
 
-import "github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/shared"
+import (
+	"encoding/json"
+
+	"github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/shared"
+)
 
 type AggregateAttestationResponse struct {
 	Data *shared.Attestation `json:"data"`
@@ -28,4 +32,12 @@ type GetAttestationDataResponse struct {
 
 type ProduceSyncCommitteeContributionResponse struct {
 	Data *shared.SyncCommitteeContribution `json:"data"`
+}
+
+// ProduceBlockV3Response is a wrapper json object for the returned block from the ProduceBlockV3 endpoint
+type ProduceBlockV3Response struct {
+	Version                 string          `json:"version"`
+	ExecutionPayloadBlinded bool            `json:"execution_payload_blinded"`
+	ExecutionPayloadValue   string          `json:"execution_payload_value"`
+	Data                    json.RawMessage `json:"data"` // represents the block values based on the version
 }
